+++ modules/__init__.py
@@ -1,3 +1,6 @@
+# coding: utf-8
+#! python3  # noqa: E265
+
 from .ui import *
 from .utils import *
 from .threads import *
